USER - 내 주소 관리

+++ src/main/java/kr/co/jhta/app/delideli/user/account/domain/UserAddress.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class UserAddress {
-    private int userAddressKey;
+    private Long userAddressKey;
     private Long userKey;
     private String address;
     private String zipcode;
